requirements.txt
streamlit
google-genai
chromadb
pypdf
python-docx
python-dotenv
watchdog # A "guard dog" that automatically refreshes the app instantly when code or data files are modified.
